Apple and Orange
Given the value of for apples and oranges, can you determine how many apples and oranges will
fall on Sam's house (i.e., in the inclusive range )? Print the number of apples that fall on Sam's house
as your first line of output, then print the number of oranges that fall on Sam's house as your second line
of output.