Birthday Cake Candles
You are in-charge of the cake for your niece's birthday and have decided the cake will have one candle for
each year of her total age. When she blows out the candles, she’ll only be able to blow out the tallest
ones.
For example, if your niece is turning years old, and the cake will have candles of height , , , , she
will be able to blow out candles successfully, since the tallest candle is of height and there are such
candles.
Complete the function Given the height of each individual candle, find and print the number of candles
she can successfully blow out.