Plus Minus
Given an array of integers, calculate the fractions of its elements that are positive, negative, and are
zeros. Pr
int the decimal value of each fraction on a new line.
Complete the plusMinus function in the editor below. It should print out the ratio of positive, negative and
zero items in the array, each on a separate line rounded to six decimals.